% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders.R
\name{add_component}
\alias{add_component}
\title{Add a component to a causal workflow}
\usage{
add_component(x, component_id, component, stage = 1L)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{component_id}{A character string identifying the component.}

\item{component}{A \code{tidymodels} object, such as a \verb{<workflow>} or
\verb{<workflow_set>}.}

\item{stage}{A single integer for the analysis stage. Defaults to \code{1L}.}
}
\value{
An updated \code{causal_workflow} object.
}
\description{
Add a component to a causal workflow
}
