% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{add_options}
\alias{add_options}
\title{Add options to a causal workflow component}
\usage{
add_options(x, component_id, options, stage = 1L)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{component_id}{A character string identifying the component.}

\item{options}{A \code{control} object, such as one created by
\code{\link[=control_propensity]{control_propensity()}}.}

\item{stage}{A single integer for the analysis stage. Defaults to \code{1L}.}
}
\value{
An updated \code{causal_workflow} object.
}
\description{
This function allows you to add a \code{control} object to a specific component
within a \code{causal_workflow}.
}
