% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_workflows.R
\name{add_propensity_model}
\alias{add_propensity_model}
\title{Add a propensity model to a causal workflow}
\usage{
add_propensity_model(x, wflow)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{wflow}{A \code{tidymodels} \code{workflow} object for the propensity model.}
}
\value{
An updated \code{causal_workflow} object.
}
\description{
Add a propensity model to a causal workflow
}
