% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders.R
\name{add_stage}
\alias{add_stage}
\title{Add a Stage to a Causal Workflow}
\usage{
add_stage(x, ..., .stage)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named arguments where the
name defines the \code{component_id} and the value is a \verb{<workflow>},
\verb{<workflow_set>}, or \verb{<tailor>} object.}

\item{.stage}{A single integer for the new stage.}
}
\value{
An updated \code{causal_workflow} object.
}
\description{
Appends a new set of components for a subsequent analysis stage to an
existing \code{causal_workflow} object.
}
