% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards-estimation.R
\name{backwards_estimation_step}
\alias{backwards_estimation_step}
\title{Perform a Single Stage of a Backwards Estimation Procedure}
\usage{
backwards_estimation_step(
  object,
  data,
  k,
  next_stage_model = NULL,
  discount = 1,
  actions
)
}
\arguments{
\item{object}{A \code{staged_workflow} object.}

\item{data}{A data frame containing all necessary variables. It must include
columns for \code{stage}, \code{action}, and \code{outcome}.}

\item{k}{The integer index of the stage to fit.}

\item{next_stage_model}{The fitted model from stage \code{k+1}. If \code{NULL} (the
default), the model for stage \code{k} is assumed to be the final stage and is
fit to the observed outcome.}

\item{discount}{A numeric value between 0 and 1 for discounting future
outcomes. Defaults to 1 (no discounting).}

\item{actions}{A character vector of all possible action levels.}
}
\value{
A fitted \code{workflow} or \code{causal_workflow} for stage \code{k}.
}
\description{
This function executes a single step of the backwards estimation algorithm
used to fit a \code{staged_workflow}. It fits the model for a specified stage \code{k},
using the fitted model from stage \code{k+1} to calculate a pseudo-outcome if
necessary.

Its primary purpose is for debugging multi-stage \code{causal_workflow} and
\code{staged_workflow}s.
}
