% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_workflows.R
\name{causal_workflow}
\alias{causal_workflow}
\title{Create a Causal Workflow}
\usage{
causal_workflow(..., .stage = 1L)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named arguments where the
name defines the \code{component_id} and the value is a \verb{<workflow>},
\verb{<workflow_set>}, or \verb{<tailor>} object.}

\item{.stage}{A single integer defining the analysis stage for the
provided components. Defaults to \code{1L}.}
}
\value{
A \code{causal_workflow} object.
}
\description{
Initializes a \code{causal_workflow} object. This function creates a tibble-based
structure that serves as a blueprint for a causal analysis. Components are
added to this object using \code{\link[=add_component]{add_component()}} or \code{\link[=add_stage]{add_stage()}}.
}
