% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalworkflows-package.R
\docType{package}
\name{causalworkflows-package}
\alias{causalworkflows}
\alias{causalworkflows-package}
\title{causalworkflows: A Framework for Multi-Component and Staged Causal Estimators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a framework for estimating causal effects using multi-stage, modular modeling procedures. The package introduces the 'causal_workflow' and 'staged_workflow' objects, which allow for flexible specification of nuisance models (e.g., for propensity scores and outcome models) and the final causal model. It leverages cross-fitting to mitigate the risks of over-fitting when constructing estimators like the Augmented Inverse Propensity Weighting (AIPW) estimator. This structure is designed to be extensible to other multi-stage estimators, such as those used in Q-learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsperger/causalworkflows}
  \item Report bugs at \url{https://github.com/jsperger/causalworkflows/issues}
}

}
\author{
\strong{Maintainer}: Sperger John \email{josp@duck.com}

}
\keyword{internal}
