% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{collect_metrics}
\alias{collect_metrics}
\title{Collect metrics from a fitted causal workflow}
\usage{
collect_metrics(x, ...)
}
\arguments{
\item{x}{A \code{fitted_causal_workflow} object.}

\item{...}{Not currently used.}
}
\value{
A tibble with the calculated causal estimates. The tibble will have
the following columns:
\itemize{
\item \code{.metric}: The name of the metric (e.g., "ate").
\item \code{.estimator}: The name of the causal estimator used (e.g., "aipw").
\item \code{.estimate}: The numeric value of the estimate.
}
}
\description{
Collect metrics from a fitted causal workflow
}
