% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_causal}
\alias{control_causal}
\alias{control_causal_grid}
\alias{control_causal_resamples}
\alias{control_causal_bayes}
\title{Control wrappers}
\usage{
control_causal_grid()

control_causal_resamples()

control_causal_bayes()
}
\value{
A \link[tune:control_grid]{tune::control_grid}, \link[tune:control_bayes]{tune::control_bayes},
or \link[tune:control_grid]{tune::control_resamples} object.
}
\description{
Supply these light wrappers as the \code{control} argument in a
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}
call to return the needed elements for use in cross-fitting for causal
inference.

These functions will return the appropriate control grid to ensure that
assessment set predictions and information on model specifications and
preprocessors, is supplied in the resampling results object!

To integrate stack settings with your existing control settings, note
that these functions just call the appropriate \verb{tune::control_*} function
with the arguments \verb{save_pred = TRUE, save_workflow = TRUE}.
}
\examples{
library(tune)

# these are the same!
control_causal_grid()
control_grid(save_pred = TRUE, save_workflow = TRUE)

}
\seealso{
See \code{example_data} for examples of these functions used in context.
}
