% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{control_causal_workflow}
\alias{control_causal_workflow}
\alias{control_propensity}
\alias{control_outcome}
\alias{control_targeting}
\title{Control options for causal workflow components}
\usage{
control_propensity(truncate = 0.01, ...)

control_outcome(control_tune = tune::control_grid(), ...)

control_targeting(fluctuation = "logistic", ...)
}
\arguments{
\item{truncate}{A numeric value between 0 and 1 for truncating propensity
scores.}

\item{...}{Not currently used.}

\item{control_tune}{A \code{tune} control object, such as \code{\link[tune:control_grid]{tune::control_grid()}},
used when a component is a \verb{<workflow_set>}.}

\item{fluctuation}{The type of fluctuation model to use in TMLE. Currently,
only \code{"logistic"} is supported.}
}
\value{
A \code{causal_workflow_control} object.
}
\description{
These functions create classed lists of options that can be supplied to the
\code{options} argument of \code{\link[=add_options]{add_options()}} to control how different components
of a \code{causal_workflow} are handled during fitting.
}
