% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_fit}
\alias{control_fit}
\title{Control settings for the recursive fitting engine}
\usage{
control_fit()
}
\value{
A \code{control_fit} object, which is a list containing control
parameters.
}
\description{
This function provides a basic control object for the \code{fit_recursive} engine.
}
