% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{twoStageCont}
\title{Toy Two-Stage Trial Datasets}
\format{
twoStageCont is a dataset generated to mimic a simple two-stage
trial. It is taken from the \code{DTRreg} package for concordance testing. The data.frame contains 1000 observations with 5 columns:
\describe{
\item{X1}{The first stage covariate. A normally distributed continuous variable.}
\item{A1}{The first stage treatment. A binary variable.}
\item{X2}{The second stage covariate. A normally distributed continuous variable.}
\item{A2}{The second stage treatment. A binary variable.}
\item{Y}{The outcome. A continuous variable.}
}
}
\usage{
data(twoStageCont)
}
\description{
These datasets are provided only to facilitate examples. They are not based
on or representative of any real-world applications.
}
\keyword{datasets}
