% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.calculate_eif_pom}
\alias{.calculate_eif_pom}
\title{Calculate the EIF for the Potential Outcome Mean (POM)}
\usage{
.calculate_eif_pom(data, treatment_var, outcome_var, treatment_levels)
}
\arguments{
\item{data}{A data frame with outcome, treatment, and nuisance predictions.}

\item{treatment_var}{A character string of the treatment variable name.}

\item{outcome_var}{A character string of the outcome variable name.}

\item{treatment_levels}{A character vector of the treatment levels.}
}
\value{
A tibble of observation-level EIF values.
}
\description{
Calculate the EIF for the Potential Outcome Mean (POM)
}
\keyword{internal}
