% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.get_nuisance_preds}
\alias{.get_nuisance_preds}
\title{Get nuisance model predictions}
\usage{
.get_nuisance_preds(g_fit, q_fit, data, treatment_var, treatment_levels)
}
\arguments{
\item{g_fit}{A fitted propensity score model.}

\item{q_fit}{A fitted outcome model.}

\item{data}{The data to predict on.}

\item{treatment_var}{A character string of the treatment variable name.}

\item{treatment_levels}{A character vector of the treatment levels.}
}
\value{
A tibble of nuisance predictions.
}
\description{
Get nuisance model predictions
}
\keyword{internal}
