% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-workflow.R
\name{.get_observed_prob}
\alias{.get_observed_prob}
\title{Helper to extract the predicted probability for the observed action}
\usage{
.get_observed_prob(data, model_prefix, observed_actions)
}
\description{
Helper to extract the predicted probability for the observed action
}
\keyword{internal}
