% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract_preprocessor.causal_workflow}
\alias{extract_preprocessor.causal_workflow}
\title{Extract a preprocessor from a causal workflow}
\usage{
\method{extract_preprocessor}{causal_workflow}(x, component_id, stage = 1L, ...)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{component_id}{A character string identifying the component from which
to extract the preprocessor.}

\item{stage}{A single integer for the analysis stage. Defaults to \code{1L}.}

\item{...}{Not used.}
}
\value{
The preprocessor object, typically a \code{recipe}.
}
\description{
Extract a preprocessor from a causal workflow
}
