% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-across.R
\name{fit_across}
\alias{fit_across}
\title{Fit a causal workflow with cross-fitting}
\usage{
fit_across(object, ...)
}
\arguments{
\item{object}{A \code{causal_workflow} object that has been configured with
a propensity model and an outcome model.}

\item{...}{Not used.}

\item{data}{A data frame containing the training data, including the
treatment, outcome, and covariate variables.}
}
\value{
A \code{fitted_causal_workflow} object. This object contains:
\itemize{
\item \code{propensity_model_fit}, \code{outcome_model_fit}: The final nuisance models
fitted on the full dataset.
\item \code{treatment_levels}: A character vector of the treatment levels.
\item \code{estimates}: A tibble with the estimated potential outcome for each
treatment level.
\item \code{variances}: A tibble with the variance of the potential outcome
estimator for each level.
\item \code{eif_pom}: A tibble of the observation-level efficient influence function
(EIF) values for the POM, with one column for each treatment level.
\item \code{nuisance_predictions}: A tibble of the out-of-sample nuisance
predictions from the cross-fitting procedure.
}
}
\description{
\code{\link[=fit_across]{fit_across()}} for a \code{causal_workflow} object performs a cross-fitted estimation
of causal effects. It uses out-of-sample predictions for the nuisance models
(propensity and outcome models) to construct the efficient influence
function (EIF) for the potential outcome mean (POM), providing robust
estimates of potential outcomes.
}
\details{
Cross-fitting (also known as cross-estimation) is a sample-splitting
technique that mitigates the bias that can arise from using the same data to
fit and predict with a model, which is particularly important when using
flexible or machine learning-based nuisance models. The data is split into
K folds. For each fold, the propensity and outcome models are fitted on the
other K-1 folds (the analysis set) and then used to generate predictions for
the held-out fold (the assessment set). This results in a full set of
out-of-sample predictions for all observations in the original dataset.

These out-of-sample predictions are then used to calculate the EIF for the
Potential Outcome Mean (POM) for each treatment level, leading to more
robust and less biased estimates of the causal effect.
}
\seealso{
\code{\link[=fit.causal_workflow]{fit.causal_workflow()}}, \code{\link[=tune_nested]{tune_nested()}}
}
