% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-staged-workflow.R
\name{multi_predict.fitted_staged_workflow}
\alias{multi_predict.fitted_staged_workflow}
\title{Predict a sequence of optimal actions}
\usage{
\method{multi_predict}{fitted_staged_workflow}(object, new_data, ...)
}
\arguments{
\item{object}{A \code{fitted_staged_workflow} object.}

\item{new_data}{A data frame containing the initial state for the first
stage of prediction.}

\item{...}{Not used.}
}
\value{
A tibble with one row per observation in \code{new_data} and columns for the
predicted action and value at each stage (e.g., \code{.pred_action_1},
\code{.pred_value_1}, etc.).
}
\description{
This function predicts the optimal action and value for each stage in a
\code{fitted_staged_workflow}.
}
\details{
This method iteratively predicts the optimal action for each stage, updates
the state based on that action, and passes the new state to the next stage's
model. This is a "greedy" prediction that assumes the optimal action is taken
at each step.
}
