% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-staged-workflow.R
\name{predict.fitted_staged_workflow}
\alias{predict.fitted_staged_workflow}
\title{Predict from a \code{fitted_staged_workflow}}
\usage{
\method{predict}{fitted_staged_workflow}(object, new_data, stage, type = "action", ...)
}
\arguments{
\item{object}{A \code{fitted_staged_workflow} object.}

\item{new_data}{A data frame containing the predictor values for the
specified stage.}

\item{stage}{An integer specifying the stage for which to generate
predictions.}

\item{type}{A character string indicating the type of prediction.
\itemize{
\item \code{"action"}: Returns the action with the maximum predicted Q-value.
\item \code{"value"}: Returns the maximum predicted Q-value.
}}

\item{...}{Not used.}
}
\value{
A tibble with a \code{.pred} column containing the predictions.
}
\description{
This function generates predictions for a specific stage from a fitted
\code{staged_workflow} object.
}
\details{
The method retrieves the fitted model for the specified \code{stage} and predicts
the Q-value for each possible action. Based on the \code{type} argument, it
returns either the optimal action or the corresponding value.
}
