% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-resamples.R, R/fit.R, R/predict.R,
%   R/tune-nested.R, R/tune.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit_resamples}
\alias{fit}
\alias{augment}
\alias{tune_grid}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{fit}}}

  \item{tune}{\code{\link[tune]{fit_resamples}}, \code{\link[tune]{tune_grid}}, \code{\link[tune]{tune_grid}}}
}}
