% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged-workflow.R
\name{set_action_exclusions}
\alias{set_action_exclusions}
\title{Set action exclusion rules for a \code{staged_workflow}}
\usage{
set_action_exclusions(x, formula)
}
\arguments{
\item{x}{A \code{staged_workflow} object.}

\item{formula}{A formula describing the exclusion rules. The LHS should be
empty, and the RHS should be a logical expression.}
}
\value{
An updated \code{staged_workflow} object.
}
\description{
This verb defines rules for actions that are not available in certain states.
}
\details{
The formula is used during the fitting process to identify valid actions
when calculating pseudo-outcomes. For example, \code{~ action == "chemo" & tumor_size > 5} would exclude the "chemo" action for patients with a tumor
size greater than 5.

This is currently a placeholder and will be used in a future version.
}
\examples{
\dontrun{
q_spec <- staged_workflow() |>
  set_action_exclusions(~ action == "chemo" & tumor_size > 5)
}
}
