% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-workflow.R
\name{tmle_workflow}
\alias{tmle_workflow}
\title{Initialize a TMLE Workflow}
\usage{
tmle_workflow(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{tmle_workflow} object, which also inherits from \code{causal_workflow}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{tmle_workflow()} function initializes a \code{tmle_workflow} object.
This object is a specialized type of \code{causal_workflow} that signals to the
fitting engine that TMLE-specific steps (e.g., model targeting or
fluctuation) are required.

Like a \code{causal_workflow}, it serves as a container for the propensity and
outcome models, which are added with \code{\link[=add_propensity_model]{add_propensity_model()}} and
\code{\link[=add_outcome_model]{add_outcome_model()}}.
}
