% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_grid.causal_workflow}
\alias{tune_grid.causal_workflow}
\title{Tune a causal workflow component}
\usage{
\method{tune_grid}{causal_workflow}(object, resamples, grid, ...)
}
\arguments{
\item{object}{A \code{\link[=causal_workflow]{causal_workflow()}} object.}

\item{resamples}{An \code{rsample} object, such as one created by \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}.}

\item{grid}{A data frame of tuning hyperparameters, created by a \code{dials}
grid function.}

\item{...}{Additional arguments passed to \code{\link[tune:tune_grid]{tune::tune_grid()}}.}
}
\value{
A \code{tune_results} object from the underlying call to \code{\link[tune:tune_grid]{tune::tune_grid()}}.
The user is responsible for using this result to finalize the appropriate
workflow and update the \code{\link[=causal_workflow]{causal_workflow()}} object.
}
\description{
This method supports "greedy" hyperparameter tuning for a single component
of a \code{\link[=causal_workflow]{causal_workflow()}}. It identifies which component (propensity or outcome
model) the provided \code{grid} of hyperparameters belongs to and dispatches
the tuning job to \code{\link[tune:tune_grid]{tune::tune_grid()}} for that component.
}
