% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-nested.R
\name{tune_nested}
\alias{tune_nested}
\title{Tune and estimate a causal workflow with nested resampling}
\usage{
tune_nested(object, resamples, ..., inner_v = 5, metric = NULL)
}
\arguments{
\item{object}{A \code{\link[=causal_workflow]{causal_workflow()}} object.}

\item{resamples}{An \code{rsample} object for the outer folds of nested
resampling, such as one created by \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}.}

\item{...}{Additional arguments passed to the underlying tuning or fitting
functions.}

\item{inner_v}{The number of folds for the inner resampling loop.}

\item{metric}{A character string for the metric to optimize during tuning.}
}
\value{
A \code{fitted_causal_workflow} object.
}
\description{
\code{\link[=tune_nested]{tune_nested()}} for a \code{\link[=causal_workflow]{causal_workflow()}} object performs nested resampling to
either tune the hyperparameters of the nuisance models or to fit an ensemble
of models using \code{stacks}. It provides robust, unbiased causal effect
estimates for complex, regularized, or non-parametric models that require
tuning.
}
\details{
Nested resampling is essential when a modeling process involves both
hyperparameter tuning and estimation of a final causal effect. It consists
of two loops:
\itemize{
\item An \strong{outer loop} that splits the data into folds for final estimation.
The models are never trained on the assessment sets of these folds,
ensuring the final estimates are unbiased.
\item An \strong{inner loop} that uses the analysis sets from the outer loop to
perform hyperparameter tuning or ensemble fitting.
}

The best-performing model or the fitted ensemble from the inner loop is then
used to generate out-of-sample predictions for the outer loop's assessment
set. These predictions are used to calculate the final, doubly robust causal
effect estimates.
}
