% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_workflows.R
\name{add_outcome_model}
\alias{add_outcome_model}
\title{Add an outcome model to a causal workflow}
\usage{
add_outcome_model(x, spec)
}
\arguments{
\item{x}{A \code{causal_workflow} object.}

\item{spec}{A \code{tidymodels} \code{workflow} or \code{workflow_set} object for the
outcome model.}
}
\value{
An updated \code{causal_workflow} object.
}
\description{
Add an outcome model to a causal workflow
}
