% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_workflows.R
\name{causal_workflow}
\alias{causal_workflow}
\title{Initialize a Causal Workflow}
\usage{
causal_workflow(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{causal_workflow} object.
}
\description{
The \code{causal_workflow()} function initializes a \code{causal_workflow} object.
This object serves as a container for the various component models
required for a causal analysis, such as a propensity score model and an
outcome model.

The resulting \code{causal_workflow} object is then passed to "verb" functions
like \code{add_propensity_model()} and \code{add_outcome_model()} to specify the
analysis.
}
\details{
This initial version of \code{causal_workflow} is designed for single-stage,
point-treatment estimators with binary treatments (e.g., AIPW). The
underlying architecture is designed to be extensible to more complex
scenarios in future versions, including:
\itemize{
\item Multi-level categorical or continuous treatments.
\item More generic weighting schemes, such as inverse probability of censoring
weights.
\item Multi-stage dynamic treatment regimes.
}
}
