% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.causal_workflow}
\alias{fit.causal_workflow}
\title{Fit a causal workflow}
\usage{
\method{fit}{causal_workflow}(object, data, ...)
}
\arguments{
\item{object}{A \code{causal_workflow} object that has been configured with
a propensity model and an outcome model.}

\item{data}{A data frame containing the training data. When used inside
\code{\link[=fit.staged_workflow]{fit.staged_workflow()}}, the response variable is expected to be in a
column named \code{outcome}.}

\item{...}{Not used.}
}
\value{
A \code{fitted_causal_workflow} object.
}
\description{
\code{fit} for a \code{causal_workflow} object performs a straightforward,
non-cross-fitted estimation of causal effects. It fits the propensity and
outcome models on the full dataset and uses in-sample predictions to
construct the efficient influence function (EIF) for the potential outcome
mean (POM).

This method is best suited for simple, non-regularized models (e.g.,
unpenalized logistic or linear regression) where the risk of overfitting is
low. For more complex or regularized models, using in-sample predictions can
lead to biased estimates. In those cases, it is strongly recommended to use
\code{\link[=fit_across]{fit_across()}} for cross-fitted estimation or \code{\link[=tune_nested]{tune_nested()}} for
estimation with hyperparameter tuning.

When called within \code{fit.staged_workflow()}, this function will be fitted
using a pseudo-outcome.
}
\seealso{
\code{\link[=fit_across]{fit_across()}}, \code{\link[=tune_nested]{tune_nested()}}, \code{\link[=fit.staged_workflow]{fit.staged_workflow()}}
}
