% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/generics.R, R/predict.R, R/resample.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{extract_preprocessor}
\alias{augment}
\alias{fit_resamples}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{fit}}}

  \item{hardhat}{\code{\link[hardhat:hardhat-extract]{extract_preprocessor}}}

  \item{tune}{\code{\link[tune]{fit_resamples}}}
}}

