% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged-workflow.R
\name{staged_workflow}
\alias{staged_workflow}
\title{\code{staged_workflow}}
\usage{
staged_workflow()
}
\value{
A new \code{\link[=staged_workflow]{staged_workflow()}} object.
}
\description{
Initializes a \code{staged_workflow} object.
}
\details{
A \code{staged_workflow} is a container for a sequence of stage-specific
\link{{.pkg tidymodels}} workflows, designed for fitting multi-stage models like
those used in Dynamic Treatment Regimes (DTRs).

The core components are:
\itemize{
\item \code{stages}: A list mapping each stage to a \code{\link[workflows:workflow]{workflows::workflow()}}.
\item \code{exclusions}: A formula specifying conditions under which certain
actions are unavailable.
}
}
\examples{
\dontrun{
library(parsnip)
library(recipes)
library(workflows)

# Define a simple linear model workflow
lm_wflow <- workflow() |>
  add_model(linear_reg()) |>
  add_formula(outcome ~ covar1 + covar2 + action)

# Initialize a staged workflow and add the model for two stages
q_spec <- staged_workflow() |>
  add_stage_model(lm_wflow, stages = 1:2)
}
}
